requirements.txt
Django==5.1.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.0
openai==1.65.5
python-dotenv==1.0.1
mysqlclient==2.2.7
django-cors-headers==4.7.0
decouple==0.0.7
requests==2.32.3